+++ jwk.test.js
@@ -5,6 +5,7 @@ import {
 	generateJWK, base64EncodeJWK, importJWKFromBase64, importRawKey, createJWKFile,
 	loadJWKFromBlob, fetchJWK, importJWK, exportJWK, exportAsRFC7517JWK,
 	importRFC7517JWK,
+	getKid,
 } from './jwk.js';
 import { createJWT, decodeToken, verifyJWT, verifySignature } from './jwt.js';
 import { ES256, HS256 } from './consts.js';
@@ -12,6 +13,18 @@ import { signal } from './signal.test.js';
 import { ALGOS } from './consts.js';
 
 const algos = Object.keys(ALGOS);
+
+const knownKey = {
+	jwk: {
+		kty: 'RSA',
+		n: '0vx7agoebGcQSuuPiLJXZptN9nndrQmbXEps2aiAFbWhM78LhWx4cbbfAAtVT86zwu1RK7aPFFxuhDR1L6tSoc_BJECPebWKRXjBZCiFV4n3oknjhMstn64tZ_2W-5JsGY4Hc5n9yBXArwl93lqt7_RN5w6Cf0h4QyQ5v-65YGjQR0_FDW2QvzqY368QQMicAtaSqzs8KJZgnYb9c7d0zgdAZHzu6qMQvRL5hajrn1n91CbOpbISD08qNLyrdkt-bFTWhAI4vMQFh6WeZu0fM4lFd2NcRwr3XPksINHaQ-G_xBniIqbw0Ls1jF44-csFCur-kEgU8awapJzKnqDKgw',
+		e: 'AQAB',
+		alg: 'RS256',
+		ext: true,
+		key_ops: ['verify'],
+	},
+	kid: 'NzbLsXh8uDCcd-6MNwXF4W_7noWXFZAfHkxZsRGC9Xs',
+};
 
 // Make concurrent due to extensive async keygen operations
 describe('Test JWK/key functions', { concurrency: true }, async () => {
@@ -36,6 +49,16 @@ describe('Test JWK/key functions', { concurrency: true }, async () => {
 			}
 		});
 	}
+
+	test('Test JWK generating `kid`s', { signal }, async () => {
+		const publicKey = await importJWK(knownKey.jwk);
+		if (publicKey instanceof Error) {
+			assert.fail(publicKey);
+		} else {
+			const kid = await getKid(publicKey);
+			assert.strictEqual(kid, knownKey.kid, '`kid` for known key should match.');
+		}
+	});
 
 	test('Import symmetric key from raw', { signal }, async () => {
 		const key = await importRawKey(crypto.getRandomValues(new Uint8Array(256)), { algorithm: HS256 });